requirements.txt
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.20.0
Pillow>=9.0.0
PyMuPDF>=1.22.0
numpy>=1.24.0
flask>=2.3.0
tqdm>=4.65.0
